refactor: move getFilteredProducts from FilterService to ProductsService
