+++ src/app/core/store/content/parameters/parameters.effects.ts
@@ -3,7 +3,7 @@ import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { concatMap, mergeMap } from 'rxjs/operators';
 
 import { Product } from 'ish-core/models/product/product.model';
-import { FilterService } from 'ish-core/services/filter/filter.service';
+import { ProductsService } from 'ish-core/services/products/products.service';
 import { loadProductSuccess } from 'ish-core/store/shopping/products';
 import { mapErrorToAction, mapToPayload } from 'ish-core/utils/operators';
 
@@ -15,14 +15,14 @@ import {
 
 @Injectable()
 export class ParametersEffects {
-  constructor(private actions$: Actions, private filterService: FilterService) {}
+  constructor(private actions$: Actions, private productsService: ProductsService) {}
 
   loadParametersProductListFilter$ = createEffect(() =>
     this.actions$.pipe(
       ofType(loadParametersProductListFilter),
       mapToPayload(),
       concatMap(({ id, searchParameter, amount }) =>
-        this.filterService.getFilteredProducts(searchParameter, amount).pipe(
+        this.productsService.getFilteredProducts(searchParameter, amount).pipe(
           mergeMap(({ products }) => [
             ...products.map((product: Product) => loadProductSuccess({ product })),
             loadParametersProductListFilterSuccess({ id, productList: products.map(p => p.sku) }),
