+++ src/app/core/store/shopping/products/products.effects.ts
@@ -28,6 +28,7 @@ import { selectRouteParam } from 'ish-core/store/core/router';
 import { setBreadcrumbData } from 'ish-core/store/core/viewconf';
 import { personalizationStatusDetermined } from 'ish-core/store/customer/user';
 import { loadCategory } from 'ish-core/store/shopping/categories';
+import { loadProductsForFilter } from 'ish-core/store/shopping/filter';
 import { getProductListingItemsPerPage, setProductListingPages } from 'ish-core/store/shopping/product-listing';
 import { loadProductPrices } from 'ish-core/store/shopping/product-prices';
 import { HttpStatusCodeService } from 'ish-core/utils/http-status-code/http-status-code.service';
@@ -185,6 +186,45 @@ export class ProductsEffects {
             ),
           ]),
           mapErrorToAction(loadProductsForMasterFail, { masterSKU })
+        )
+      )
+    )
+  );
+
+  loadFilteredProducts$ = createEffect(() =>
+    this.actions$.pipe(
+      ofType(loadProductsForFilter),
+      mapToPayload(),
+      switchMap(({ id, searchParameter, page, sorting }) =>
+        this.store.pipe(
+          select(getProductListingItemsPerPage(id.type)),
+          whenTruthy(),
+          first(),
+          switchMap(pageSize =>
+            this.productsService
+              .getFilteredProducts(searchParameter, pageSize, sorting, ((page || 1) - 1) * pageSize)
+              .pipe(
+                mergeMap(({ products, total, sortableAttributes }) => [
+                  ...products.map((product: Product) => loadProductSuccess({ product })),
+                  setProductListingPages(
+                    this.productListingMapper.createPages(
+                      products.map(p => p.sku),
+                      id.type,
+                      id.value,
+                      pageSize,
+                      {
+                        filters: id.filters,
+                        itemCount: total,
+                        startPage: page,
+                        sortableAttributes,
+                        sorting,
+                      }
+                    )
+                  ),
+                ]),
+                mapErrorToAction(loadProductFail)
+              )
+          )
         )
       )
     )
